% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jlst.R
\name{jlsp}
\alias{jlsp}
\title{jlsp}
\usage{
jlsp(y, x, covar = NULL, covar.var = FALSE, var.type = 1)
}
\arguments{
\item{y}{vector of outcome values.}

\item{x}{vector of exposure values.}

\item{covar}{a data.frame of covariates.}

\item{covar.var}{adjust the second stage (variance component) of the approach by the covariates.}

\item{type}{type of test (default: 1 [Breusch-Pagan variance test]; options: 1 [Breusch-Pagan variance test], 2 [Brown-Forsythe variance test]).}
}
\value{
a list of results. Q/F is the test statistic, DF is the degrees of freedom and P is the p-value. The model coefficients from each part of the model are given in the coef objects.
}
\description{
jlsp performs the joint location-and-scale test using Fisher's method.
}
\examples{
x <- rbinom(1000, 1, 0.5)
y <- 0.5 + rnorm(1000, 0.025, 0.025)*x + rnorm(1000, 0, 0.1)
jlsp(y, x, type=2)
}
\author{
James R Staley <jrstaley95@gmail.com>
}
