% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jlst.R
\name{jlssc}
\alias{jlssc}
\title{jlssc}
\usage{
jlssc(y, x, covar = NULL, type = 1)
}
\arguments{
\item{y}{vector of outcome values.}

\item{x}{vector of exposure values.}

\item{covar}{a data.frame of covariates.}

\item{type}{type of test (default: 1 [Breusch-Pagan variance test]; options: 1 [Breusch-Pagan variance test], 2 [Brown-Forsythe variance test], 3 [Method of moments version of test 1], 4 [Method of moments version of test 2]).}
}
\value{
a data.frame of results. Q is the test statistic, DF is the degrees of freedom and P is the p-value.
}
\description{
jlssc performs the joint location-and-scale score test.
}
\examples{
x <- rbinom(1000, 1, 0.5)
y <- 0.5 + rnorm(1000, 0.025, 0.025)*x + rnorm(1000, 0, 0.1)
jlssc(y, x)
}
\author{
James R Staley <james.staley@bristol.ac.uk>
}
